pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.90.0",
    "pydub>=0.25.1",
    "pytelegrambotapi>=4.27.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
